rtrb

Module chunks

Writing and reading multiple items at once into and from a RingBuffer.

Multiple items at once can be moved from an iterator into the ring buffer by using Producer::write_chunk_uninit() followed by WriteChunkUninit::fill_from_iter(). Alternatively, mutable access to the (uninitialized) slots of the chunk can be obtained with WriteChunkUninit::as_mut_slices(), which requires writing some unsafe code. To avoid that, Producer::write_chunk() can be used, which initializes all slots with their Default value and provides mutable access by means of WriteChunk::as_mut_slices().

Multiple items at once can be moved out of the ring buffer by using Consumer::read_chunk() and iterating over the returned ReadChunk (or by explicitly calling ReadChunk::into_iter()). Immutable access to the slots of the chunk can be obtained with ReadChunk::as_slices().

If the item type T implements Copy, the convenience functions Producer::push_partial_slice(), Consumer::pop_partial_slice() and Consumer::pop_partial_slice_uninit() can be used.

Examples

The following examples use a single thread for simplicity, but in a real application, producer and consumer would of course live on different threads:

If the trait bound T: Copy is satisfied, Producer::push_partial_slice() and Consumer::pop_partial_slice() (and Consumer::pop_partial_slice_uninit()) can be used.

use rtrb::RingBuffer;

let (mut producer, mut consumer) = RingBuffer::new(4);

let source = vec![1, 2, 3, 4, 5, 6];
let (pushed, remainder) = producer.push_partial_slice(&source);
assert_eq!(pushed, [1, 2, 3, 4]);
assert_eq!(remainder, [5, 6]);

let mut destination = vec![0; 3];
let (popped, remainder) = consumer.pop_partial_slice(&mut destination);
assert_eq!(popped, [1, 2, 3]);
assert_eq!(remainder, []);
assert_eq!(destination, [1, 2, 3]);

let (popped, remainder) = consumer.pop_partial_slice(&mut destination);
assert_eq!(popped, [4]);
assert_eq!(remainder, [2, 3]);
// The returned slices are mutable sub-slices into `destination`.
remainder[0] = 99;
assert_eq!(destination, [4, 99, 3]);

If this convenience interface is too limited (or if T is not Copy) the more fundamental methods Producer::write_chunk(), Producer::write_chunk_uninit() and Consumer::read_chunk() can be used.

use rtrb::RingBuffer;

let (mut producer, mut consumer) = RingBuffer::new(5);

if let Ok(chunk) = producer.write_chunk_uninit(4) {
    chunk.fill_from_iter([10, 11, 12]);
    // Note that we requested 4 slots but we've only written to 3 of them!
} else {
    unreachable!();
}

assert_eq!(producer.slots(), 2);
assert_eq!(consumer.slots(), 3);

if let Ok(chunk) = consumer.read_chunk(2) {
    assert_eq!(chunk.into_iter().collect::<Vec<_>>(), [10, 11]);
} else {
    unreachable!();
}

// One element is still in the queue:
assert_eq!(consumer.peek(), Ok(&12));

let data = vec![20, 21, 22, 23];
producer.push_entire_slice(&data).unwrap();

assert!(producer.is_full());
assert_eq!(consumer.slots(), 5);

let mut v = Vec::<i32>::with_capacity(5);
if let Ok(chunk) = consumer.read_chunk(5) {
    let (first, second) = chunk.as_slices();
    v.extend(first);
    v.extend(second);
    chunk.commit_all();
} else {
    unreachable!();
}
assert_eq!(v, [12, 20, 21, 22, 23]);
assert!(consumer.is_empty());

The iterator API can be used to move items from one ring buffer to another:

use rtrb::{Consumer, Producer};

fn move_items<T>(src: &mut Consumer<T>, dst: &mut Producer<T>) -> usize {
    let n = src.slots().min(dst.slots());
    dst.write_chunk_uninit(n).unwrap().fill_from_iter(src.read_chunk(n).unwrap())
}

Write as many slots as possible, given an iterator (and return the number of written slots):

use rtrb::{Producer, chunks::ChunkError::TooFewSlots};

fn push_from_iter<T, I>(queue: &mut Producer<T>, iter: I) -> usize
where
    T: Default,
    I: IntoIterator<Item = T>,
{
    let iter = iter.into_iter();
    let n = match iter.size_hint() {
        (_, None) => queue.slots(),
        (_, Some(n)) => n,
    };
    let chunk = match queue.write_chunk_uninit(n) {
        Ok(chunk) => chunk,
        // Remaining slots are returned, this will always succeed:
        Err(TooFewSlots(n)) => queue.write_chunk_uninit(n).unwrap(),
    };
    chunk.fill_from_iter(iter)
}

Structs

ReadChunk

Structure for reading from multiple slots in one go.

ReadChunkIntoIter

An iterator that moves out of a ReadChunk.

WriteChunk

Structure for writing into multiple (Default-initialized) slots in one go.

WriteChunkUninit

Structure for writing into multiple (uninitialized) slots in one go.

Enums

ChunkError

Error type for Consumer::read_chunk(), Consumer::pop_entire_slice(), Consumer::pop_entire_slice_uninit(), Producer::write_chunk(), Producer::write_chunk_uninit() and Producer::push_entire_slice().