rtrb::chunks
Struct WriteChunk

pub struct WriteChunk<'a, T>(/* private fields */);

Structure for writing into multiple (Default-initialized) slots in one go.

This is returned from Producer::write_chunk().

To obtain uninitialized slots, use Producer::write_chunk_uninit() instead, which also allows moving items from an iterator into the ring buffer by means of WriteChunkUninit::fill_from_iter().

Implementations

impl<T> WriteChunk<'_, T>
where T: Default,

pub fn as_mut_slices(&mut self) -> (&mut [T], &mut [T])

Returns two slices for writing to the requested slots.

All slots are initially filled with their Default value.

The first slice can only be empty if 0 slots have been requested. If the first slice contains all requested slots, the second one is empty.

After writing to the slots, they are not automatically made available to be read by the Consumer. This has to be explicitly done by calling commit() or commit_all(). If items are written but not committed afterwards, they will not become available for reading and they will be leaked (which is only relevant if T implements Drop).

pub fn commit(self, n: usize)

Makes the first n slots of the chunk available for reading.

The rest of the chunk is dropped.

Panics

Panics if n is greater than the number of slots in the chunk.

pub fn commit_all(self)

Makes the whole chunk available for reading.

pub fn len(&self) -> usize

Returns the number of slots in the chunk.

pub fn is_empty(&self) -> bool

Returns true if the chunk contains no slots.